requirements.txt
# Core dependencies
pyrogram==2.0.106
TgCrypto==1.2.5
aiohttp==3.9.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Additional features
aiofiles==23.2.1
humanize==4.9.0
Pillow==10.1.0
qrcode==7.4.2
matplotlib==3.8.2

# Monitoring and logging
colorlog==6.8.0
sentry-sdk==1.39.1

# Development tools
black==23.12.0
flake8==6.1.0
pytest==7.4.3
pytest-asyncio==0.23.2